Monet/requirements.txt
datasets
qwen_vl_utils
vllm==0.10.0
mathruler
torch
pylatexenc
torchvision
Pillow
transformers==4.54.0
trl==0.15.2
tqdm
safetensors
accelerate
deepspeed
ray
einops
sentencepiece
requests
numpy
scipy
opencv-python-headless
fastapi
uvicorn
pydantic
modelscope
wandb